+++ .phan/config.php
@@ -654,6 +654,8 @@ return [
         'RemoveDebugStatementPlugin',
         'UnsafeCodePlugin',
         'DeprecateAliasPlugin',
+        // Suggest '@return never'
+        '.phan/plugins/AddNeverReturnTypePlugin.php',
         // Still have false positives to suppress
         // '.phan/plugins/StaticVariableMisusePlugin.php',