+++ internal/lib/IncompatibleRealStubsSignatureDetector.php
@@ -60,7 +60,6 @@ class IncompatibleRealStubsSignatureDetector extends IncompatibleSignatureDetect
         if (!is_string($realpath)) {
             echo "Could not find realpath of '$dir'\n";
             static::printUsageAndExit();
-            return;
         }
         $this->directory = $realpath;