+++ src/Phan/Language/Type/FunctionLikeDeclarationType.php
@@ -131,9 +131,12 @@ abstract class FunctionLikeDeclarationType extends Type implements FunctionInter
         });
     }
 
+    /**
+     * @return never
+     */
     public function __clone()
     {
-        throw new \AssertionError('Should not clone ClosureTypeDeclaration');
+        throw new \AssertionError('Should not clone ' . static::class);
     }
 
     /**
@@ -389,6 +392,7 @@ abstract class FunctionLikeDeclarationType extends Type implements FunctionInter
 
     /**
      * @unused-param $fqsen
+     * @return never
      * @override
      */
     public function setFQSEN(FQSEN $fqsen): void
@@ -417,6 +421,7 @@ abstract class FunctionLikeDeclarationType extends Type implements FunctionInter
 
     /**
      * @override
+     * @return never
      * @param list<Node|int|string|float> $args
      */
     public function analyzeFunctionCall(CodeBase $code_base, Context $context, array $args, Node $node = null): void
@@ -428,6 +433,7 @@ abstract class FunctionLikeDeclarationType extends Type implements FunctionInter
      * @unused-param $context
      * @unused-param $code_base
      * @param Parameter[] $parameter_list @unused-param
+     * @return never
      * @override
      */
     public function analyzeWithNewParams(Context $context, CodeBase $code_base, array $parameter_list): Context
@@ -438,19 +444,26 @@ abstract class FunctionLikeDeclarationType extends Type implements FunctionInter
     /**
      * @override
      * @unused-param $parameter
+     * @return never
      */
     public function appendParameter(Parameter $parameter): void
     {
         throw new \AssertionError('unexpected call to ' . __METHOD__);
     }
 
-    /** @override */
+    /**
+     * @override
+     * @return never
+     */
     public function clearParameterList(): void
     {
         throw new \AssertionError('unexpected call to ' . __METHOD__);
     }
 
-    /** @override */
+    /**
+     * @override
+     * @return never
+     */
     public function cloneParameterList(): void
     {
         throw new \AssertionError('unexpected call to ' . __METHOD__);
@@ -485,6 +498,7 @@ abstract class FunctionLikeDeclarationType extends Type implements FunctionInter
     /**
      * @override
      * @param list<Node|int|string|float> $args
+     * @return never
      */
     public function getDependentReturnType(CodeBase $code_base, Context $context, array $args): UnionType
     {
@@ -535,7 +549,10 @@ abstract class FunctionLikeDeclarationType extends Type implements FunctionInter
         return true;
     }
 
-    /** @override */
+    /**
+     * @override
+     * @return never
+     */
     public function getInternalScope(): ClosedScope
     {
         throw new \AssertionError('unexpected call to ' . __METHOD__);
@@ -638,6 +655,7 @@ abstract class FunctionLikeDeclarationType extends Type implements FunctionInter
 
     /**
      * @param list<Parameter> $parameter_list
+     * @return never
      */
     public function setParameterList(array $parameter_list): void
     {
@@ -646,6 +664,7 @@ abstract class FunctionLikeDeclarationType extends Type implements FunctionInter
 
     /**
      * @internal - moves real parameter defaults to the inferred phpdoc parameters
+     * @return never
      */
     public function inheritRealParameterDefaults(): void
     {
@@ -697,6 +716,9 @@ abstract class FunctionLikeDeclarationType extends Type implements FunctionInter
         return false;
     }
 
+    /**
+     * @return never
+     */
     public function recordOutputReferenceParamName(string $parameter_name): void
     {
         throw new \AssertionError('unexpected call to ' . __METHOD__);
@@ -709,22 +731,32 @@ abstract class FunctionLikeDeclarationType extends Type implements FunctionInter
 
     /**
      * @unused
+     * @return never
      */
     public function setComment(Comment $comment): void
     {
         throw new \AssertionError('unexpected call to ' . __METHOD__);
     }
 
+    /**
+     * @return never
+     */
     public function setFunctionCallAnalyzer(Closure $analyzer): void
     {
         throw new \AssertionError('unexpected call to ' . __METHOD__);
     }
 
+    /**
+     * @return never
+     */
     public function addFunctionCallAnalyzer(Closure $analyzer): void
     {
         throw new \AssertionError('unexpected call to ' . __METHOD__);
     }
 
+    /**
+     * @return never
+     */
     public function setDependentReturnTypeClosure(Closure $analyzer): void
     {
         throw new \AssertionError('unexpected call to ' . __METHOD__);
@@ -733,6 +765,7 @@ abstract class FunctionLikeDeclarationType extends Type implements FunctionInter
     /**
      * @override
      * @unused-param $has_return
+     * @return never
      */
     public function setHasReturn(bool $has_return): void
     {
@@ -742,6 +775,7 @@ abstract class FunctionLikeDeclarationType extends Type implements FunctionInter
     /**
      * @override
      * @unused-param $has_yield
+     * @return never
      */
     public function setHasYield(bool $has_yield): void
     {
@@ -756,6 +790,7 @@ abstract class FunctionLikeDeclarationType extends Type implements FunctionInter
     /**
      * @override
      * @unused-param $has_static_variable
+     * @return never
      */
     public function setHasStaticVariable(bool $has_static_variable): void
     {
@@ -767,6 +802,9 @@ abstract class FunctionLikeDeclarationType extends Type implements FunctionInter
         return false;
     }
 
+    /**
+     * @return never
+     */
     public function setInternalScope(ClosedScope $scope): void
     {
         throw new \AssertionError('unexpected call to ' . __METHOD__);
@@ -775,6 +813,7 @@ abstract class FunctionLikeDeclarationType extends Type implements FunctionInter
     /**
      * @unused-param $is_return_type_undefined
      * @override
+     * @return never
      */
     public function setIsReturnTypeUndefined(bool $is_return_type_undefined): void
     {
@@ -784,6 +823,7 @@ abstract class FunctionLikeDeclarationType extends Type implements FunctionInter
     /**
      * @unused-param $number
      * @override
+     * @return never
      */
     public function setNumberOfOptionalParameters(int $number): void
     {
@@ -793,12 +833,16 @@ abstract class FunctionLikeDeclarationType extends Type implements FunctionInter
     /**
      * @unused-param $number
      * @override
+     * @return never
      */
     public function setNumberOfRequiredParameters(int $number): void
     {
         throw new \AssertionError('unexpected call to ' . __METHOD__);
     }
 
+    /**
+     * @return never
+     */
     public function setPHPDocParameterTypeMap(array $parameter_map): void
     {
         throw new \AssertionError('unexpected call to ' . __METHOD__);
@@ -806,6 +850,7 @@ abstract class FunctionLikeDeclarationType extends Type implements FunctionInter
 
     /**
      * @param ?UnionType $union_type the raw phpdoc union type
+     * @return never
      */
     public function setPHPDocReturnType(?UnionType $union_type): void
     {
@@ -888,12 +933,16 @@ abstract class FunctionLikeDeclarationType extends Type implements FunctionInter
 
     /**
      * @param list<string> $suppress_issue_list
+     * @return never
      */
     public function setSuppressIssueSet(array $suppress_issue_list): void
     {
         throw new \AssertionError('should not call ' . __METHOD__);
     }
 
+    /**
+     * @return never
+     */
     public function setUnionType(UnionType $type): void
     {
         throw new \AssertionError('should not call ' . __METHOD__);
@@ -1102,6 +1151,9 @@ abstract class FunctionLikeDeclarationType extends Type implements FunctionInter
         return !$other->isDefiniteNonCallableType($code_base);
     }
 
+    /**
+     * @return never
+     */
     public function setAttributeList(array $attribute_list): void
     {
         throw new \AssertionError('should not call ' . __METHOD__);